Move package_cache to its solve consumer
