+++ conda_build/conda_interface.py
@@ -17,6 +17,7 @@ from conda.auxlib.packaging import (  # noqa: F401
 from conda.base.context import context, determine_target_prefix, reset_context
 from conda.base.context import non_x86_machines as non_x86_linux_machines  # noqa: F401
 from conda.core.package_cache import ProgressiveFetchExtract  # noqa: F401
+from conda.core.package_cache_data import PackageCacheData  # noqa: F401
 from conda.exceptions import (  # noqa: F401
     CondaError,
     CondaHTTPError,
@@ -185,5 +186,4 @@ from ._legacy_conda_imports import (
     install_actions,
     linked,
     linked_data,
-    package_cache,
 )
